*: private and public filestore
All media is handled in the media path defined in settings.py !
The idea is to have only once the file content kept in the
filestore (_filestore directory). So, if a user uploads multiple
times the same file, it will only take once the disk space.
File reference are saved in `FileReference` model. When a file
is uploaded, its path will be a symlink from the public/private
filestore to the real filestore (_filestore). The real file is
removed once there no reference to the file, otherwise, it is
kept.

Public file:
Public file are stored in the filestore, and their symlink is
created from 'public' directory. Symlink is defined by the
`upload_to` field attribute. File are accessible directly on
/media/public/<path created with upload_to>/. This public subdir
is serve by nginx directly for performance sake.

Private file:
Private file are stored in the filestore, and their symlink is
created from 'private' directory. Symlink is defined by the
`upload_to` field attribute.
Private file are accessible on /media/private/<path created with
upload_to>?signature=<signature>&expires=1730838944
The url is signed for one hour; if the user can access the resource,
he can access for one hour to its related file.
A special nginx config is required to make it work: the `internal`
directive ! Django will check the signed url and redirect the
request to the non exposed real filestore path, made accessible
with `internal` nginx directive.