user: access rules
This commit gives the base tool to implement access
rules in business apps.

A rule is a Q-expression applied on a model for list of
actions. This is implemented as python class,
hardcoded to be more efficient for now (no need to
have a too big degree of customization).

The rule class must then be 'register' on a user role,
to link them. A global rule is register on no role !
the evaluation is odoo-like: logical AND for global
rules combined (and) with logical OR on role rules.

Note: Current registration can be replaced by identifying
rule class with a unique key, and attach to role in the
database, like for permissions.